osm-verifier/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx[asyncio]>=0.27.0
pydantic>=2.7.0
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
dateparser>=1.2.0
sentence-transformers>=3.0.1
scikit-learn>=1.4.2
pillow>=10.3.0
scikit-image>=0.23.2
numpy>=1.26.4
requests>=2.31.0
aiohttp>=3.9.5
brave-search==0.1.8